Created 'push' functionality for conveyor belt
You can now pass in an array, cell array or individual items to the conveyor by calling the "push" function on the conveyor to push these items along it. "push" checks if items are properly placed on top of the conveyor before pushing.
Created a test script to demo the "push" function