Update E05 worker and related test file
"searchpickUP" now returns a joint path instead of a final joint position. search Pickup now requires an initial joint position to be given. These changes were made so that the inverse kinematic calculators could find more optimal paths. plan PickupPath has been changed to now use the paths returned from "searchPickupPath". An addition, more logging messages have been added to functions to show user when the function is resorting to default values for arguments. Changed 'liftUp' function to "searchLiftUp' and made it take in a q0 for same reasons as 'searchPickUp'. Added function 'planDepositPath'. Updated 'animateArm' to now takse in an item thats in the gripper so that it can animated to move with the arm.